signal: move into tokio-net
