+++ ci/patch.toml
@@ -9,7 +9,6 @@ tokio-fs = { path = "tokio-fs" }
 tokio-io = { path = "tokio-io" }
 tokio-macros = { path = "tokio-macros" }
 tokio-net = { path = "tokio-net" }
-tokio-signal = { path = "tokio-signal" }
 tokio-sync = { path = "tokio-sync" }
 tokio-timer = { path = "tokio-timer" }
 tokio-tls = { path = "tokio-tls" }
