+++ tokio-net/Cargo.toml
@@ -22,6 +22,12 @@ categories = ["asynchronous", "network-programming"]
 
 [features]
 async-traits = []
+signal = [
+    "futures-util-preview",
+    "mio-uds",
+    "libc",
+    "signal-hook-registry",
+]
 tcp = [
     "bytes",
     "futures-util-preview",
@@ -62,10 +68,11 @@ futures-sink-preview = { version = "=0.3.0-alpha.18", optional = true }
 futures-util-preview = { version = "=0.3.0-alpha.18", optional = true }
 iovec = { version = "0.1", optional = true }
 
-# UDS
 [target.'cfg(unix)'.dependencies]
+# UDS / Signal
 mio-uds = { version = "0.6.5", optional = true }
 libc = { version = "0.2.42", optional = true }
+signal-hook-registry = { version = "~1", optional = true }
 
 [dev-dependencies]
 tokio = { version = "=0.2.0-alpha.1", path = "../tokio" }
