+++ tokio-net/src/signal/ctrl_c.rs
@@ -1,9 +1,8 @@
 #[cfg(unix)]
-use crate::unix::Signal as Inner;
+use super::unix::Signal as Inner;
 #[cfg(windows)]
-use crate::windows::Event as Inner;
-
-use tokio_net::driver::Handle;
+use super::windows::Event as Inner;
+use crate::driver::Handle;
 
 use futures_core::stream::Stream;
 use std::io;
