+++ tokio-net/src/signal/windows.rs
@@ -7,9 +7,9 @@
 
 #![cfg(windows)]
 
-use crate::registry::{globals, EventId, EventInfo, Init, Storage};
+use crate::driver::Handle;
+use super::registry::{globals, EventId, EventInfo, Init, Storage};
 
-use tokio_net::driver::Handle;
 use tokio_sync::mpsc::{channel, Receiver};
 
 use futures_core::stream::Stream;
