+++ tokio-net/tests/signal_dropping_does_not_deregister_other_instances.rs
@@ -1,9 +1,9 @@
 #![cfg(unix)]
+#![cfg(feature = "signal")]
 #![warn(rust_2018_idioms)]
 #![feature(async_await)]
 
-use libc;
-
+#[path = "support/signal.rs"]
 pub mod support;
 use crate::support::*;
