+++ tokio-net/tests/signal_multi_loop.rs
@@ -1,10 +1,11 @@
 #![cfg(unix)]
+#![cfg(feature = "signal")]
 #![warn(rust_2018_idioms)]
 
+#[path = "support/signal.rs"]
 pub mod support;
 use crate::support::*;
 
-use libc;
 use std::sync::mpsc::channel;
 use std::thread;
