+++ tokio-net/tests/support/signal.rs
@@ -1,16 +1,16 @@
 #![cfg(unix)]
 #![warn(rust_2018_idioms)]
 
+use tokio::timer::Timeout;
+pub use tokio::runtime::current_thread::{self, Runtime as CurrentThreadRuntime};
+pub use tokio_net::signal::unix::{Signal, SignalKind};
+
 use futures_util::future::FutureExt;
 use libc::{c_int, getpid, kill};
 use std::future::Future;
 use std::time::Duration;
-use tokio_timer::Timeout;
-
 pub use futures_util::future;
 pub use futures_util::stream::StreamExt;
-pub use tokio::runtime::current_thread::{self, Runtime as CurrentThreadRuntime};
-pub use tokio_signal::unix::{Signal, SignalKind};
 
 pub fn with_timeout<F: Future>(future: F) -> impl Future<Output = F::Output> {
     Timeout::new(future, Duration::from_secs(1)).map(Result::unwrap)
