--- tokio-signal/Cargo.toml
@@ -1,43 +1,0 @@
-[package]
-name = "tokio-signal"
-# When releasing to crates.io:
-# - Remove path dependencies
-# - Update html_root_url.
-# - Update doc url
-#   - Cargo.toml
-# - Update CHANGELOG.md.
-# - Create "v0.3.x" git tag.
-version = "0.3.0-alpha.1"
-edition = "2018"
-authors = ["Tokio Contributors <team@tokio.rs>"]
-license = "MIT"
-repository = "https://github.com/tokio-rs/tokio"
-homepage = "https://github.com/tokio-rs/tokio"
-documentation = "https://docs.rs/tokio-signal/0.3.0-alpha.1/tokio_signal"
-description = """
-An implementation of an asynchronous Unix signal handling backed futures.
-"""
-categories = ["asynchronous"]
-
-[dependencies]
-futures-core-preview = "=0.3.0-alpha.18"
-futures-util-preview = "=0.3.0-alpha.18"
-lazy_static = "1"
-tokio-io = { version = "=0.2.0-alpha.1", path = "../tokio-io" }
-tokio-net = { version = "=0.2.0-alpha.1", path = "../tokio-net" }
-tokio-sync = { version = "=0.2.0-alpha.1", path = "../tokio-sync" }
-
-[target.'cfg(unix)'.dependencies]
-libc = "0.2"
-mio = "0.6.14"
-mio-uds = "0.6"
-signal-hook-registry = "~1"
-
-[dev-dependencies]
-tokio = { version = "=0.2.0-alpha.1", path = "../tokio" }
-tokio-timer = { version = "=0.3.0-alpha.1", path = "../tokio-timer" }
-tokio-sync = { version = "=0.2.0-alpha.1", path = "../tokio-sync", features = ["async-traits"]}
-
-[target.'cfg(windows)'.dependencies.winapi]
-version = "0.3"
-features = ["consoleapi", "minwindef", "wincon"]
